Remove obsolete profile page
