+++ lib/constants.ts
@@ -1,3 +1,11 @@
 export const OBJECTIVES = ["Ahorrar", "Invertir", "Salir de deudas", "Otro"];
 export const LEARNING_PREFERENCES = ["Videos", "Juegos", "Lecturas", "Simuladores"];
 export const INTERESTS = ["Cripto", "Inversión", "Economía básica", "VC", "Impuestos", "Deuda"];
+
+import { Grid3X3, MessageCircle, Heart } from "lucide-react";
+
+export const PROFILE_TABS = [
+  { id: "grid", icon: Grid3X3 },
+  { id: "comments", icon: MessageCircle },
+  { id: "favorites", icon: Heart },
+];
