+++ lib/types.ts
@@ -1,10 +1,10 @@
 export interface FormData {
-  name: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-  age: number;
-  objectives: string[];
-  learningPreferences: string[];
-  interests: string[];
+  name?: string;
+  email?: string;
+  password?: string;
+  confirmPassword?: string;
+  age?: number;
+  objectives?: string[];
+  learningPreferences?: string[];
+  interests?: string[];
 }
