+++ package.json
@@ -9,6 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "react-icons": "^5.0.1"
+    "@hookform/resolvers": "^3.3.2",
+    "lucide-react": "^0.294.0",
+    "react-hook-form": "^7.49.2",
+    "react-icons": "^5.0.1",
+    "zod": "^3.22.4"
   }
 }
